import refactoring part of common/compiler (#19120)
